requirements.txt
Django==4.1
django_heroku==0.3.1
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-dotenv==1.4.2
django-heroku==0.3.1
fontawesomefree==6.1.2
gunicorn==20.1.0
pip==22.1.2
pydotplus==2.0.2
six==1.16.0
django-filter==22.1
Pillow==9.2.0
black==22.8.0